Made build.fsx run against newer FAKE
Recently FAKE made a breaking change. Rather than locking `fszmq` to a specific version of FAKE, code has been updated to work with the latest version. If there are build errors, delete `FAKE` from the `packages` directory are try again (this will force the bootstrapper to fetch a newer version of `FAKE`).